requirements.txt
django
django-crontab
djangorestframework
django-cors-headers
djangorestframework-simplejwt